+++ nahieran-js-doc/jsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "baseUrl": "src",
+    "checkJs": true,
+    "jsx": "react"
+  },
+  "include": ["src"]
+}
