+++ nahieran-js-doc/package.json
@@ -16,7 +16,8 @@
     "react-markdown": "^8.0.4",
     "react-scripts": "5.0.1",
     "react-syntax-highlighter": "^15.5.0",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^2.1.4",
+    "nahieran-js": "file:../src"
   },
   "scripts": {
     "start": "react-scripts start",
