+++ nahieran-js-doc/src/App.js
@@ -1,9 +1,7 @@
-// import './App.css';
 import '@fontsource/roboto/300.css';
 import '@fontsource/roboto/400.css';
 import '@fontsource/roboto/500.css';
 import '@fontsource/roboto/700.css';
-import { useState } from "react";
 import MenuAppBar from './Components/AppBar';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
 import Container from '@mui/material/Container';
@@ -12,12 +10,16 @@ import Toolbar from '@mui/material/Toolbar';
 import CssBaseline from '@mui/material/CssBaseline';
 import {
   Card, CardContent, Typography, CardActions, Button,
-  Dialog, DialogTitle, DialogContent,
-  DialogContentText, DialogActions
 } from "@mui/material"
-import { getTVPrograms } from "http://127.0.0.1:3001/src/index.js";
 import MarkDown from './Components/MarkDown';
 import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
+// import { getTVPrograms } from "http://127.0.0.1:3001/src/index.js";
+import { getTVPrograms } from "nahieran-js";
+import React from 'react';
+import DocCard from './Components/DocCard';
+
+
+
 
 
 
@@ -33,30 +35,7 @@ function App() {
   });
 
 
-  const [openDialog, setOpenDialog] = useState(false);
-  const [dialogTitle, setDialogTitle] = useState("");
-  const [dialogContent, setDialogContent] = useState("");
-
-
-  const handleClose = () => {
-    setOpenDialog(false);
-  };
-
-
-
-  const getTVProgramsHandler = () => {
-
-    setOpenDialog(true);
-    setDialogTitle("Get TV program list");
-    getTVPrograms().then(res => {
-      console.log("OK: ", res);
-      setDialogContent(res);
-
-    });
-  }
-
   const markdown = `
-
   ~~~js
   import { getTVPrograms } from "nahieran-js";
 
@@ -79,34 +58,34 @@ function App() {
 
 
 
-          <Card variant="outlined" sx={{ marginBottom: 2 }}>
-            <CardContent>
-              <Typography variant="h3" color="text.secondary" gutterBottom>
-                Getting started
-              </Typography>
-              <MarkDown markdown={`
-  This project is based on [@erral's](https://github.com/erral) [https://github.com/erral/eitbapi-react](https://github.com/erral/eitbapi-react) project.
 
-  A set of JS functions to use the EITB Nahieran API
-  ## Install 
-  ~~~bash
-  npm i nahieran-js
-  ~~~
-  `}
-              />
-            </CardContent>
-          </Card>
+          <DocCard
+            title="Getting started"
+            markdown={`
+This project is based on [@erral's](https://github.com/erral) [https://github.com/erral/eitbapi-react](https://github.com/erral/eitbapi-react) project.
 
+A set of JS functions to use the EITB Nahieran API
+## Install 
+~~~bash
+npm i nahieran-js
+~~~
+            `} />
 
-          <Typography variant="h2" id="TV" color="text.secondary" gutterBottom>
+          <Typography variant="h2" id="TV" color="text.secondary" gutterBottom sx={{ marginTop: 3 }}>
             TV
           </Typography>
-          <Card variant="outlined">
+
+
+          <DocCard
+            title="Get TV program list"
+            markdown={markdown}
+            demoFunction={getTVPrograms} />
+          {/* <Card variant="outlined">
 
 
             <CardContent>
               <Typography variant="h3" color="text.secondary" gutterBottom>
-                Get TV program list
+                
               </Typography>
 
 
@@ -115,31 +94,9 @@ function App() {
             <CardActions>
               <Button size="small" onClick={getTVProgramsHandler}>Try it!</Button>
             </CardActions>
-          </Card>
+          </Card> */}
 
-          <Dialog
-            open={openDialog}
-            onClose={handleClose}
-            aria-labelledby="alert-dialog-title"
-            aria-describedby="alert-dialog-description"
-          >
-            <DialogTitle id="alert-dialog-title">
-              {dialogTitle}
-            </DialogTitle>
-            <DialogContent>
-              <DialogContentText id="alert-dialog-description">
 
-                {dialogContent &&
-                  JSON.stringify(dialogContent)}
-                {/* // <SyntaxHighlighter language="json"></SyntaxHighlighter> */}
-              </DialogContentText>
-            </DialogContent>
-            <DialogActions>
-              <Button onClick={handleClose} autoFocus>
-                Close
-              </Button>
-            </DialogActions>
-          </Dialog>
 
 
         </Box>
